% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_matrx_by_year.R
\name{cor_matrx_by_year}
\alias{cor_matrx_by_year}
\title{Correlation Matrix of Numerical NBA Stats by Year}
\usage{
cor_matrx_by_year(year)
}
\arguments{
\item{year}{A number for years between 1950 and 2017.}
}
\value{
double type of correlation matrix
}
\description{
This function returns a correlation matrix of the numerical NBA stats from a specified year between 1950 and 2017.
}
\examples{
cor_matrx_by_year()
}
\keyword{Correlation}
\keyword{NBA}
\keyword{by}
\keyword{matrix}
\keyword{year}
