% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_by_year.R
\name{oldest_by_year}
\alias{oldest_by_year}
\title{Oldest NBA Player's Age in a given Year}
\usage{
oldest_by_year(year)
}
\arguments{
\item{year}{A number for years between 1950 and 2017.}
}
\value{
list row with Year, Player, and Age
}
\description{
This function returns the age of the oldest (limited to year) player  of the NBA according to specified year.
}
\examples{
oldest_by_year()
}
\keyword{NBA}
\keyword{Oldest}
