% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_rndm_age.R
\name{year_rndm_age}
\alias{year_rndm_age}
\title{Random NBA Player's Age by Year}
\usage{
year_rndm_age(year)
}
\arguments{
\item{year}{A number for years between 1950 and 2017.}
}
\value{
list row with Year, Player, and Age
}
\description{
This function returns the age of a random player of the NBA according to specified year.
}
\examples{
year_rndm_age()
}
\keyword{Age}
\keyword{NBA}
\keyword{by}
\keyword{year}
